+++ src/lightning_utilities/core/rank_zero.py
@@ -5,9 +5,9 @@
 """Utilities that can be used for calling functions on a particular rank."""
 
 import logging
+import sys
 import warnings
 from functools import wraps
-from platform import python_version
 from typing import Any, Callable, Optional, TypeVar, Union
 
 from typing_extensions import ParamSpec, overload
@@ -46,7 +46,7 @@ def rank_zero_only(fn: Callable[P, T], default: Optional[T] = None) -> Callable[
 
 
 def _debug(*args: Any, stacklevel: int = 2, **kwargs: Any) -> None:
-    if python_version() >= "3.8.0":
+    if sys.version_info >= (3, 8, 0):  # noqa: UP036
         kwargs["stacklevel"] = stacklevel
     log.debug(*args, **kwargs)
 
@@ -58,7 +58,7 @@ def rank_zero_debug(*args: Any, stacklevel: int = 4, **kwargs: Any) -> None:
 
 
 def _info(*args: Any, stacklevel: int = 2, **kwargs: Any) -> None:
-    if python_version() >= "3.8.0":
+    if sys.version_info >= (3, 8, 0):  # noqa: UP036
         kwargs["stacklevel"] = stacklevel
     log.info(*args, **kwargs)