+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased] - 2025-MM-DD
+
+### Added
+
+### Changed
+
+### Fixed
+
+- Fixed python version parsing for logging `stacklevel` ([#375](https://github.com/Lightning-AI/utilities/pull/375))
+
+
+---
+
 ## [0.14.0] - 2025-03-07
 
 ### Added